使用LongAdder优化多线程并发下的计数器自增原子操作，提高性能

+++ src/main/java/com/fangxuele/tool/push/logic/KeFuPriorMsgServiceThread.java
@@ -86,7 +86,7 @@ public class KeFuPriorMsgServiceThread extends BaseMsgServiceThread {
             tableModel.setValueAt((int) ((double) (i + 1) / list.size() * 100), tableRow, 5);
 
             // 总进度条
-            MainWindow.mainWindow.getPushTotalProgressBar().setValue((int) (PushData.successRecords + PushData.failRecords));
+            MainWindow.mainWindow.getPushTotalProgressBar().setValue(PushData.successRecords.intValue() + PushData.failRecords.intValue());
         }
 
         // 当前线程结束
